apps/sink/requirements.txt
kafka-python==2.0.2
requests==2.32.3
orjson==3.10.7
